Remove AWS specific command and several enhancements
Removed `AwsPublishCommand` because it's context-specific and not appropriate for the boilerplate generator. Updated the `composer.json` to reflect accurate PHP version compatibility and include `laravel/prompts` for better user interactivity. Used Laravel prompts in Domain, Test, and Install commands to standardize and simplify interactive prompts. Updated description and readme for better clarity. In addition, console output is improved for successful package installation and other operations.